Permission checks added to mainactivity, could be moved out of acitivty class to dedicated permission class or similar -- but this seems fine for the moment.
booleans indicating prescence of internet and fine location are created in application level provider to avoid having to use the system level permissions checkers everytime one wants to check gps etc.

TODO: likely should quit out of app if internet is denied or otherwise insist to user that this app is useless without internet. Addtional fiddling with snackbars on permission checking would be a good look to improve user experience.